+++ nix/ci/pie.nix
@@ -9,10 +9,7 @@
   libstemmer,
   src,
 }: let
-  pie = fetchurl {
-    url = "https://github.com/php/pie/releases/download/1.4.10/pie.phar";
-    sha256 = "b88792235c8e80be568436d4cb043b49fd1869c89b64e83d23e2882ae19d70a8";
-  };
+  pie = import ../pie.nix {inherit fetchurl;};
 in
   stdenv.mkDerivation {
     pname = "stemmer-pie-check";