+++ php_stemmer.h
@@ -1,7 +1,7 @@
 #ifndef PHP_STEMMER_PHP_H
 #define PHP_STEMMER_PHP_H 1
 
-#define PHP_STEMMER_VERSION "2.0.1"
+#define PHP_STEMMER_VERSION "2.0.2"
 #define PHP_STEMMER_RELEASE "2026-09-12"
 #define PHP_STEMMER_LICENSE "BSD-3-Clause"
 #define PHP_STEMMER_AUTHORS "© anno Domini nostri Jesu Christi 2008-2026 Javeline B.V., John Boehr, & contributors"